split main.ts part
