+++ package.json
@@ -22,6 +22,7 @@
     "genversion": "^3.0.2",
     "get-stream": "^6.0.1",
     "lodash": "^4.17.21",
+    "mustache": "^4.2.0",
     "pdfmake": "^0.2.0",
     "zeebe-node": "^1.3.2"
   },
@@ -33,6 +34,7 @@
     "@types/debug": "^4.1.5",
     "@types/lodash": "^4.14.174",
     "@types/mocha": "^9.0.0",
+    "@types/mustache": "^4.1.2",
     "@types/pdfjs-dist": "^2.7.4",
     "@types/pdfmake": "^0.1.18",
     "chai": "^4.3.4",
